+++ lib/odata_query.dart
@@ -1,1 +1,3 @@
-export 'scr/odata_query.dart';
+library odata_query;
+
+export 'package:odata_query/src/odata_query.dart';